+++ src/basic/components/features/CartSidebar/index.ts
@@ -1,0 +1,1 @@
+export { CartSidebar } from "./CartSidebar";
